requirements.txt
Flask==2.1.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
psycopg2-binary==2.9.3
python-dotenv==0.19.2
reportlab==3.6.8